Pack and cleanup vcv gui code in a single file
